fixing warning

+++ OpenGL/Source/NPCCamera.cpp
@@ -109,34 +109,34 @@ void NPCCamera::WBmove(double dt, float xUpper, float xLower, float zUpper, floa
 
 	if (Application::IsKeyPressed('D'))
 	{
-		position += right * NPCCamera_SPEED * dt;
+		position += right * NPCCamera_SPEED * (float)(dt);
 		if (checkCollisions(xUpper, xLower,zUpper,zLower))
 		{
-			position -= right * NPCCamera_SPEED * dt;
+			position -= right * NPCCamera_SPEED * (float)(dt);
 		}
 	}
 	if (Application::IsKeyPressed('W'))
 	{
-		position += front * NPCCamera_SPEED * dt;
+		position += front * NPCCamera_SPEED * (float)(dt);
 		if (checkCollisions(xUpper, xLower, zUpper, zLower))
 		{
-			position -= front * NPCCamera_SPEED * dt;
+			position -= front * NPCCamera_SPEED * (float)(dt);
 		}
 	}
 	if (Application::IsKeyPressed('A'))
 	{
-		position -= right * NPCCamera_SPEED * dt;
+		position -= right * NPCCamera_SPEED * (float)(dt);
 		if (checkCollisions(xUpper, xLower, zUpper, zLower))
 		{
-			position += right * NPCCamera_SPEED * dt;
+			position += right * NPCCamera_SPEED * (float)(dt);
 		}
 	}
 	if (Application::IsKeyPressed('S'))
 	{
-		position -= front * NPCCamera_SPEED * dt;
+		position -= front * NPCCamera_SPEED * (float)(dt);
 		if (checkCollisions(xUpper, xLower, zUpper, zLower))
 		{
-			position += front * NPCCamera_SPEED * dt;
+			position += front * NPCCamera_SPEED * (float)(dt);
 		}
 	}
 }
