+++ OpenGL/Source/c_MultiplayerLevelTwo.cpp
@@ -386,7 +386,7 @@ void c_MultiplayerLevelTwo::Update(double dt)
 	//-------------------------------------------//
 
 	Timer += (float)dt;
-	Countdown -= (float)Timer * dt;
+	Countdown -= (float)(Timer * dt);
 	FreezeTime = (float)(dt + (dt * 0));
 
 	if (Countdown <= 0)
@@ -457,7 +457,7 @@ void c_MultiplayerLevelTwo::Update(double dt)
 	if (PoneFinish)
 	{
 		if (OelapsedTime >= 20 && OelapsedTime <= 50)
-			OelapsedTime += (dt + 2);
+			OelapsedTime += (float)(dt + 2);
 
 		if (OelapsedTime >= 61 && OelapsedTime <= 106)
 			Ponelaps = 1;
@@ -478,7 +478,7 @@ void c_MultiplayerLevelTwo::Update(double dt)
 	if (PTwoFinish)
 	{
 		if (TelapsedTime >= 20 && TelapsedTime <= 50)
-			TelapsedTime += (dt + 2);
+			TelapsedTime += (float)(dt + 2);
 
 		if (TelapsedTime >= 61 && TelapsedTime <= 106)
 			PTwolaps = 1;
@@ -1514,7 +1514,7 @@ void c_MultiplayerLevelTwo::renderEnviroment()
 
 void c_MultiplayerLevelTwo::renderRain()
 {
-	for (int i = 0; i < rain.getX().size() - 3000; i++)
+	for (int i = 0; i < (int)rain.getX().size() - 3000; i++)
 	{
 		modelStack.PushMatrix();
 		modelStack.Translate(rain.getX().at(i), rain.getY().at(i), rain.getZ().at(i));
@@ -1540,7 +1540,7 @@ void c_MultiplayerLevelTwo::renderRain()
 
 void c_MultiplayerLevelTwo::RenderSnow()
 {
-	for (int i = 0; i < snow.getX().size() - 3000; i++)
+	for (int i = 0; i < (int)snow.getX().size() - 3000; i++)
 	{
 		modelStack.PushMatrix();
 		modelStack.Translate(snow.getX().at(i), snow.getY().at(i), snow.getZ().at(i));
